ingestion/pyproject.toml
[tool.poetry]
name = "ingestion"
version = "0.1.0"
description = ""
authors = ["Alessandro Mancini <alessandro.mancini4@enel.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mongo = "^0.2.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
